// src/helpers/enum/firebase.enum.ts
export enum AndroidNotificationPriority {
  min = 'min', low = 'low', default = 'default', high = 'high', max = 'max'
}

export enum AndroidMessagePriority {
  normal = 'normal',
  high = 'high'
}

export enum AndroidNotificationVisibility {
  private = 'private', public = 'public', secret = 'secret'
}