// src/notifications/enums/notifications.enum.ts
export enum NotificationsTitle {
  PATIENT_PAYMENT = 'PATIENT_PAYMENT',
  BUCKETS = 'BUCKETS',
  RIDER_PICKED_BUCKET_FROM_PHARMACY = 'RIDER_PICKED_BUCKET_FROM_PHARMACY',
  RIDER_ARRIVED_TO_PATIENT = 'RIDER_ARRIVED_TO_PATIENT',
  PAY_YOUR_RIDER = 'PAY_YOUR_RIDER',
  BUCKET_DELIVERED = 'BUCKET_DELIVERED',
  ORDER_COMPLETED = 'ORDER_COMPLETED',
  PROCEED_WITHOUT_BUCKET = 'PROCEED_WITHOUT_BUCKET',
  NO_PHARMACY_FOUND = 'NO_PHARMACY_FOUND', // no nearby pharmacy
  ORDER_CANCELLED = 'ORDER_CANCELLED' // when no pharmacy accept any bucket
}

export enum NotificationsBody {
  PATIENT_PAYMENT_BODY = 'PATIENT_PAYMENT_BODY',
  BUCKETS_BODY = 'BUCKETS_BODY',
  RIDER_PICKED_BUCKET_FROM_PHARMACY_BODY = 'RIDER_PICKED_BUCKET_FROM_PHARMACY_BODY',
  RIDER_ARRIVED_TO_PATIENT_BODY = 'RIDER_ARRIVED_TO_PATIENT_BODY',
  PAY_YOUR_RIDER_BODY = 'PAY_YOUR_RIDER_BODY',
  BUCKET_DELIVERED_BODY = 'BUCKET_DELIVERED_BODY',
  ORDER_COMPLETED_BODY = 'ORDER_COMPLETED_BODY',
  PROCEED_WITHOUT_BUCKET_BODY = 'PROCEED_WITHOUT_BUCKET_BODY',
  NO_PHARMACY_FOUND_BODY = 'NO_PHARMACY_FOUND_BODY',
  ORDER_CANCELLED_BODY = 'ORDER_CANCELLED_BODY'
}

export enum ClickAction {
  FLUTTER_NOTIFICATION_CLICK = "FLUTTER_NOTIFICATION_CLICK",
}